remove debug stack impl
this actually has no impact on performance, because we were very
careful to make the debug stack zero-sized and no-ops in release
builds. But since you can easily create OwnedStreams and convert
them back into the inner values over and over without validating
the debug stack it's not actually useful enough to warrant all
the weirdness that it takes to wire it up. Streams using a Stack
internally will be valid.